[#178] refactor: components 폴더 구조 재구성
- components/common 폴더 삭제
- 카테고리별 폴더 생성 (button, input, popup)
- 나머지 컴포넌트는 components 바로 아래 배치
- import 경로 업데이트
- 아키텍처 문서 업데이트
