+++ src/app/record/[recordId]/page.tsx
@@ -7,13 +7,13 @@ import Image from 'next/image';
 import { RecordDetail } from '@/domains/record/components/RecordDetail';
 import { useRecordActions } from '@/domains/record/hooks/useRecordActions';
 import { AsyncBoundary } from '@/lib/async';
-import Layout from '@/components/common/Layout';
-import NavigationHeader from '@/components/common/NavigationHeader';
-import Avatar from '@/components/common/Avatar';
+import Layout from '@/components/Layout';
+import NavigationHeader from '@/components/NavigationHeader';
+import Avatar from '@/components/Avatar';
 import LevelIcon from '@/domains/place/components/LevelIcon';
-import LayerPopup from '@/components/common/LayerPopup';
+import LayerPopup from '@/components/popup/LayerPopup';
 import ReportForm from '@/domains/record/components/ReportForm';
-import Loading from '@/components/common/Loading';
+import Loading from '@/components/Loading';
 import useUserStore from '@/store/user';
 import { fromNowFormat } from '@/utils/common';
 import type { LevelColor } from '@/domains/place/types/entity';
